Add after-execution tagging support (both in client and CLI)
You can now modify tags after a run has executed. A CLI provides
the ability to see all the tags, add or remove them.

The client also allows you to operate on tags directly for all
objects.

Metadata service will need to be updated to version 2.0.6 for this
to work with the metadata service. This currently works as is with
local metadata.